requirements.txt
# Data manipulation and analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning (optional)
# tensorflow>=2.8.0
# torch>=1.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Data validation
pandas-profiling>=3.0.0

# Experiment tracking
mlflow>=1.20.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
